dump configs

+++ home/bin/update-zapret.bash
@@ -1,29 +1,39 @@
 #!/usr/bin/env bash
 
-TMP=$(mktemp -d)
-TMPZ="/tmp/zapret"
-if [[ -d "$TMPZ" ]]; then
-    rm -rf "$TMPZ"
+if ! pacman -Q libnetfilter_queue &>/dev/null; then
+    sudo pacman -S libnetfilter_queue
 fi
 
-FILES=(config youtube-hosts.txt discord-hosts.txt ipset/zapret-hosts-user-exclude.txt)
+TMPZ="/tmp/zapret"
+TMPC="/tmp/zapret-configs"
+[[ -d "$TMPZ" ]] && rm -rf "$TMPZ"
 
+if [[ ! -d /opt/zapret ]]; then
+    git clone --depth 1 https://github.com/bol-van/zapret "$TMPZ" || exit 1
+    "$TMPZ"/install_easy.sh
+    exit
+fi
+
+FILES=(youtube-hosts.txt discord-hosts.txt ipset/zapret-hosts-user-exclude.txt)
+
+install -vD "/opt/zapret/$file" "$TMPC/config"
 for file in "${FILES[@]}"; do
-    install -vD "/opt/zapret/$file" "$TMP/$file"
+    install -vD "/opt/zapret/$file" "$TMPC/$file"
 done
 
 git clone --depth 1 https://github.com/bol-van/zapret "$TMPZ" || exit 1
 
 for file in "${FILES[@]}"; do
-    install -vD "$TMP/$file" "$TMPZ/$file"
+    install -vD "$TMPC/$file" "$TMPZ/$file"
 done
 
-nvim -d "$TMPZ"/config.default "$TMPZ"/config
+nvim -d "$TMPZ"/config.default "$TMPC"/config
 
 sudo rm -rf /opt/zapret/
 
 "$TMPZ"/install_easy.sh
 
-for file in "${FILES[@]}"; do
-    sudo chown alex /opt/zapret/"$file"
-done
+(
+    cd /opt/zapret/
+    sudo chown alex "${FILES[@]}"
+)
